Fix "relationship not found" error on some saves
If the working list included a filter on relationship state, and
the state of a relationship was changed to a value not cached by
that filter, the relationship would show up as "not found" after
the new state had been saved.

With this commit, a requested relationship will always be shown,
even if it's not contained in the current working list.